+++ packages/app-data/CHANGELOG.md
@@ -1,5 +1,49 @@
 # Changelog
 
+## [5.1.3](https://github.com/cowprotocol/cow-sdk/compare/sdk-app-data-v5.1.2...sdk-app-data-v5.1.3) (2026-05-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.11.0
+  * devDependencies
+    * @cowprotocol/sdk-config bumped to 2.2.0
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.7
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.7
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.21
+
+## [5.1.2](https://github.com/cowprotocol/cow-sdk/compare/sdk-app-data-v5.1.1...sdk-app-data-v5.1.2) (2026-05-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.10.3
+  * devDependencies
+    * @cowprotocol/sdk-config bumped to 2.1.0
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.6
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.6
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.20
+
+## [5.1.1](https://github.com/cowprotocol/cow-sdk/compare/sdk-app-data-v5.1.0...sdk-app-data-v5.1.1) (2026-05-20)
+
+
+### 🐛 Bug Fixes
+
+* app data cid lazy import ([#883](https://github.com/cowprotocol/cow-sdk/issues/883)) ([fd0235b](https://github.com/cowprotocol/cow-sdk/commit/fd0235b63e957ce268ce0c93626e45a396cea06e))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.5
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.5
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.19
+
 ## [5.1.0](https://github.com/cowprotocol/cow-sdk/compare/sdk-app-data-v5.0.0...sdk-app-data-v5.1.0) (2026-04-28)