+++ packages/bridging/CHANGELOG.md
@@ -1,5 +1,75 @@
 # Changelog
 
+## [4.1.0](https://github.com/cowprotocol/cow-sdk/compare/sdk-bridging-v4.0.5...sdk-bridging-v4.1.0) (2026-05-27)
+
+
+### ✨ Features
+
+* move solana to supported chain id ([#873](https://github.com/cowprotocol/cow-sdk/issues/873)) ([da8a7c2](https://github.com/cowprotocol/cow-sdk/commit/da8a7c24cd07afcd465199844c480ebe3e81383d))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-app-data bumped to 5.1.3
+    * @cowprotocol/sdk-common bumped to 0.11.0
+    * @cowprotocol/sdk-config bumped to 2.2.0
+    * @cowprotocol/sdk-contracts-ts bumped to 3.1.0
+    * @cowprotocol/sdk-cow-shed bumped to 0.3.11
+    * @cowprotocol/sdk-order-book bumped to 3.2.0
+    * @cowprotocol/sdk-trading bumped to 2.1.0
+    * @cowprotocol/sdk-weiroll bumped to 0.1.33
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.7
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.7
+    * @cowprotocol/sdk-order-signing bumped to 1.0.4
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.21
+
+## [4.0.5](https://github.com/cowprotocol/cow-sdk/compare/sdk-bridging-v4.0.4...sdk-bridging-v4.0.5) (2026-05-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-app-data bumped to 5.1.2
+    * @cowprotocol/sdk-common bumped to 0.10.3
+    * @cowprotocol/sdk-config bumped to 2.1.0
+    * @cowprotocol/sdk-contracts-ts bumped to 3.0.3
+    * @cowprotocol/sdk-cow-shed bumped to 0.3.10
+    * @cowprotocol/sdk-order-book bumped to 3.1.0
+    * @cowprotocol/sdk-trading bumped to 2.0.5
+    * @cowprotocol/sdk-weiroll bumped to 0.1.32
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.6
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.6
+    * @cowprotocol/sdk-order-signing bumped to 1.0.3
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.20
+
+## [4.0.4](https://github.com/cowprotocol/cow-sdk/compare/sdk-bridging-v4.0.3...sdk-bridging-v4.0.4) (2026-05-20)
+
+
+### 📚 Documentation
+
+* remove info about umbrella sdk ([#874](https://github.com/cowprotocol/cow-sdk/issues/874)) ([bb0280e](https://github.com/cowprotocol/cow-sdk/commit/bb0280e7bfcdd6c2d7b122baf43910d9e4e4299f))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-app-data bumped to 5.1.1
+    * @cowprotocol/sdk-contracts-ts bumped to 3.0.2
+    * @cowprotocol/sdk-cow-shed bumped to 0.3.9
+    * @cowprotocol/sdk-trading bumped to 2.0.4
+    * @cowprotocol/sdk-weiroll bumped to 0.1.31
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.5
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.5
+    * @cowprotocol/sdk-order-signing bumped to 1.0.2
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.19
+
 ## [4.0.3](https://github.com/cowprotocol/cow-sdk/compare/sdk-bridging-v4.0.2...sdk-bridging-v4.0.3) (2026-04-28)