+++ packages/flash-loans/CHANGELOG.md
@@ -1,5 +1,66 @@
 # Changelog
 
+## [3.1.0](https://github.com/cowprotocol/cow-sdk/compare/sdk-flash-loans-v3.0.5...sdk-flash-loans-v3.1.0) (2026-05-27)
+
+
+### ✨ Features
+
+* move solana to supported chain id ([#873](https://github.com/cowprotocol/cow-sdk/issues/873)) ([da8a7c2](https://github.com/cowprotocol/cow-sdk/commit/da8a7c24cd07afcd465199844c480ebe3e81383d))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.11.0
+    * @cowprotocol/sdk-app-data bumped to 5.1.3
+    * @cowprotocol/sdk-trading bumped to 2.1.0
+    * @cowprotocol/sdk-order-signing bumped to 1.0.4
+    * @cowprotocol/sdk-order-book bumped to 3.2.0
+    * @cowprotocol/sdk-config bumped to 2.2.0
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.7
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.7
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.21
+
+## [3.0.5](https://github.com/cowprotocol/cow-sdk/compare/sdk-flash-loans-v3.0.4...sdk-flash-loans-v3.0.5) (2026-05-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.10.3
+    * @cowprotocol/sdk-app-data bumped to 5.1.2
+    * @cowprotocol/sdk-trading bumped to 2.0.5
+    * @cowprotocol/sdk-order-signing bumped to 1.0.3
+    * @cowprotocol/sdk-order-book bumped to 3.1.0
+    * @cowprotocol/sdk-config bumped to 2.1.0
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.6
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.6
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.20
+
+## [3.0.4](https://github.com/cowprotocol/cow-sdk/compare/sdk-flash-loans-v3.0.3...sdk-flash-loans-v3.0.4) (2026-05-20)
+
+
+### 🐛 Bug Fixes
+
+* **flash-loans:** simplify CollateralSwapOrder interface ([#877](https://github.com/cowprotocol/cow-sdk/issues/877)) ([0edf7d1](https://github.com/cowprotocol/cow-sdk/commit/0edf7d158a65403190de3348097a2e464e24e28d))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-app-data bumped to 5.1.1
+    * @cowprotocol/sdk-trading bumped to 2.0.4
+    * @cowprotocol/sdk-order-signing bumped to 1.0.2
+  * devDependencies
+    * @cowprotocol/sdk-ethers-v5-adapter bumped to 0.4.5
+    * @cowprotocol/sdk-ethers-v6-adapter bumped to 0.4.5
+    * @cowprotocol/sdk-viem-adapter bumped to 0.3.19
+
 ## [3.0.3](https://github.com/cowprotocol/cow-sdk/compare/sdk-flash-loans-v3.0.2...sdk-flash-loans-v3.0.3) (2026-04-28)